+++ src/main/java/me/kiriyaga/nami/feature/gui/newgui/base/NamiScreen.java
@@ -1,0 +1,12 @@
+package me.kiriyaga.nami.feature.gui.newgui.base;
+
+import net.minecraft.client.gui.screen.Screen;
+import net.minecraft.text.Text;
+
+public class NamiScreen extends Screen {
+    protected NamiScreen(Text text) {
+        super(text);
+    }
+
+    // TODO: rewrite whole fucking click gui and hud from scratch
+}
