+++ src/main/java/me/kiriyaga/nami/feature/gui/newgui/base/PanelRenderer.java
@@ -11,7 +11,7 @@ import java.awt.*;
 
 
 import static me.kiriyaga.nami.Nami.*;
-import static me.kiriyaga.nami.feature.gui.oldgui.base.GuiConstants.toRGBA;
+import static me.kiriyaga.nami.util.ColorUtils.toRGBA;
 
 
 public class PanelRenderer {
