--- src/main/java/me/kiriyaga/nami/feature/gui/oldgui/base/GuiComponent.java
@@ -1,9 +1,0 @@
-package me.kiriyaga.nami.feature.gui.oldgui.base;
-
-import net.minecraft.client.gui.DrawContext;
-import net.minecraft.client.font.TextRenderer;
-
-public interface GuiComponent {
-    void render(DrawContext context, TextRenderer textRenderer, int mouseX, int mouseY, float delta);
-    boolean mouseClicked(double mouseX, double mouseY, int button);
-}
