--- src/main/java/me/kiriyaga/nami/feature/gui/oldgui/base/NamiScreen.java
@@ -1,6 +1,0 @@
-package me.kiriyaga.nami.feature.gui.oldgui.base;
-
-public class NamiScreen {
-
-    // TODO: rewrite whole fucking click gui and hud from scratch
-}
