+++ src/main/java/me/kiriyaga/nami/feature/gui/oldgui/settings/BoolSettingRenderer.java
@@ -8,7 +8,7 @@ import net.minecraft.client.gui.DrawContext;
 import java.awt.*;
 
 import static me.kiriyaga.nami.Nami.*;
-import static me.kiriyaga.nami.feature.gui.oldgui.base.GuiConstants.*;
+import static me.kiriyaga.nami.util.ColorUtils.*;
 
 public class BoolSettingRenderer implements SettingRenderer<BoolSetting> {
