+++ src/main/java/me/kiriyaga/nami/feature/gui/oldgui/settings/KeyBindSettingRenderer.java
@@ -11,7 +11,7 @@ import org.lwjgl.glfw.GLFW;
 import java.awt.*;
 
 import static me.kiriyaga.nami.Nami.*;
-import static me.kiriyaga.nami.feature.gui.oldgui.base.GuiConstants.*;
+import static me.kiriyaga.nami.util.ColorUtils.*;
 
 public class KeyBindSettingRenderer implements SettingRenderer<KeyBindSetting> {
     private static KeyBindSetting waitingForKeyBind = null;
